modified classes.js, index.html, and added box.png to the folder assets
Modified index.html so that is references the whole Phaser library instead of just the arcade physics library.
Modified classes.js to create the player info box in the bottom left corner, as well as deleted unnecessary code and added comments to every function and unclear function call.
Added box.png to assets to be used for the player info box. This image will probably be changed in further versions to look more aesthetically pleasing/fit our color scheme, which has yet to be discussed.